[WFCORE-1618] Move ActiveOperationImpl out of AbstractMessageHandler to make AMH less complex
